pyproject.toml
[tool.poetry]
name = "ordermanager"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "~3.12"
fastapi = "~0.110.2"
uvicorn = "~0.29.0"
pydantic-settings = "^2.2.1"
gunicorn = "~22.0.0"
psycopg2-binary = "~2.9.9"
sqlalchemy = "~2.0.29"
alembic = "~1.13.1"
pydantic = {extras = ["email"], version = "^2.7.1"}

[tool.poetry.group.dev.dependencies]
factory-boy = "~3.3.0"
pre-commit = "~3.7.0"
ruff = "~0.4.1"
pytest = "~8.1.1"
gitlint = "~0.19.1"
coverage = "~7.4.4"
radon = "^6.0.1"
xenon = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
